requirements.txt
keyboard
PyAutoIt
pillow
python-imagesearch
easyocr
pyinstaller